test: migrate batch-4 (notifications + PublicFindNodes) ITs to acceptance seam
