+++ core/src/test/java/com/zextras/carbonio/files/acceptance/AddedAndRemovedNodeNotificationMoveApiIT.java
@@ -2,21 +2,14 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package com.zextras.carbonio.files.api.notifications;
+package com.zextras.carbonio.files.acceptance;
 
-import com.google.inject.Injector;
-import com.zextras.carbonio.files.Simulator;
-import com.zextras.carbonio.files.Simulator.SimulatorBuilder;
 import com.zextras.carbonio.files.TestUtils;
-import com.zextras.carbonio.files.api.utilities.DatabasePopulator;
+import com.zextras.carbonio.files.acceptance.seam.FilesTestApp;
+import com.zextras.carbonio.files.acceptance.seam.impl.GuiceNettyFilesTestAppBuilder;
 import com.zextras.carbonio.files.api.utilities.GraphqlCommandBuilder;
 import com.zextras.carbonio.files.api.utilities.entities.SimplePopulatorFolder;
-import com.zextras.carbonio.files.config.FilesConfig;
 import com.zextras.carbonio.files.dal.dao.ebean.ACL;
-import com.zextras.carbonio.files.dal.repositories.interfaces.FileVersionRepository;
-import com.zextras.carbonio.files.dal.repositories.interfaces.LinkRepository;
-import com.zextras.carbonio.files.dal.repositories.interfaces.NodeRepository;
-import com.zextras.carbonio.files.utilities.MockFilesConfig;
 import com.zextras.carbonio.files.utilities.http.HttpRequest;
 import com.zextras.carbonio.files.utilities.http.HttpResponse;
 import org.assertj.core.api.Assertions;
@@ -30,17 +23,12 @@ import java.util.Map;
 
 class AddedAndRemovedNodeNotificationMoveApiIT {
 
-  static Simulator simulator;
-  static NodeRepository nodeRepository;
-  static FileVersionRepository fileVersionRepository;
-  static LinkRepository linkRepository;
-  static MockFilesConfig mockConfig;
+  static FilesTestApp app;
 
   @BeforeAll
   static void init() {
-    simulator =
-        SimulatorBuilder.aSimulator()
-            .init()
+    app =
+        GuiceNettyFilesTestAppBuilder.aFilesTestApp()
             .withDatabase()
             .withServiceDiscover()
             .withUserManagement(
@@ -49,31 +37,25 @@ class AddedAndRemovedNodeNotificationMoveApiIT {
                     "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa",
                     "fake-token-2",
                     "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaab"))
-            .build()
-            .start();
-
-    final Injector injector = simulator.getInjector();
-    nodeRepository = injector.getInstance(NodeRepository.class);
-    fileVersionRepository = injector.getInstance(FileVersionRepository.class);
-    linkRepository = injector.getInstance(LinkRepository.class);
-    mockConfig = (MockFilesConfig) injector.getInstance(FilesConfig.class);
+            .build();
   }
 
   @AfterEach
   void cleanUp() {
-    simulator.resetDatabase();
-    simulator.reinitializeMocks();
-    mockConfig.setAreNotificationsEnabled(true);
+    app.backdoor().resetDatabase();
+    app.mocks().reset();
+    app.mocks().setNotificationsEnabled(true);
   }
 
   @AfterAll
   static void cleanUpAll() {
-    simulator.stopAll();
+    app.close();
   }
 
   private void createBaseScenario() {
     // Create base folders
-    DatabasePopulator.aNodePopulator(simulator.getInjector())
+    app.backdoor()
+        .populator()
         .addNode(
             new SimplePopulatorFolder(
                 "00000000-0000-0000-0000-000000000000", "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa", "folder"))
@@ -93,7 +75,7 @@ class AddedAndRemovedNodeNotificationMoveApiIT {
     HttpRequest httpRequest =
         HttpRequest.of("POST", "/graphql/", "ZM_AUTH_TOKEN=fake-token", bodyPayload);
 
-    TestUtils.sendRequest(httpRequest, simulator.getNettyChannel());
+    app.send(httpRequest);
 
     bodyPayload =
         GraphqlCommandBuilder.aMutationBuilder("createShare")
@@ -106,7 +88,7 @@ class AddedAndRemovedNodeNotificationMoveApiIT {
     httpRequest =
         HttpRequest.of("POST", "/graphql/", "ZM_AUTH_TOKEN=fake-token", bodyPayload);
 
-    TestUtils.sendRequest(httpRequest, simulator.getNettyChannel());
+    app.send(httpRequest);
 
     // Create a node on shared directory
     bodyPayload =
@@ -119,7 +101,7 @@ class AddedAndRemovedNodeNotificationMoveApiIT {
     httpRequest =
         HttpRequest.of("POST", "/graphql/", "ZM_AUTH_TOKEN=fake-token", bodyPayload);
 
-    HttpResponse httpResponse = TestUtils.sendRequest(httpRequest, simulator.getNettyChannel());
+    HttpResponse httpResponse = app.send(httpRequest);
 
     Map<String, Object> page =
         TestUtils.jsonResponseToMap(httpResponse.getBodyPayload(), "createFolder");
@@ -137,7 +119,7 @@ class AddedAndRemovedNodeNotificationMoveApiIT {
     httpRequest =
         HttpRequest.of("POST", "/graphql/", "ZM_AUTH_TOKEN=fake-token", bodyPayload);
 
-    TestUtils.sendRequest(httpRequest, simulator.getNettyChannel());
+    app.send(httpRequest);
   }
 
   @Test
@@ -155,8 +137,7 @@ class AddedAndRemovedNodeNotificationMoveApiIT {
         HttpRequest.of("POST", "/graphql/", "ZM_AUTH_TOKEN=fake-token-2", bodyPayload);
 
     // When
-    HttpResponse httpResponse =
-        TestUtils.sendRequest(httpRequest, simulator.getNettyChannel());
+    HttpResponse httpResponse = app.send(httpRequest);
 
     // Then
     Assertions.assertThat(httpResponse.getStatus()).isEqualTo(200);
@@ -174,11 +155,7 @@ class AddedAndRemovedNodeNotificationMoveApiIT {
   @Test
   void givenANodeMovedFromOneSharedDirectoryToAnotherSharedDirectoryAndDisabledNotificationsNoNotificationShouldBeSavedOrReturned() {
     // Given
-    ((MockFilesConfig)
-        simulator
-            .getInjector()
-            .getInstance(FilesConfig.class))
-        .setAreNotificationsEnabled(false);
+    app.mocks().setNotificationsEnabled(false);
 
     createBaseScenario();
 
@@ -192,8 +169,7 @@ class AddedAndRemovedNodeNotificationMoveApiIT {
         HttpRequest.of("POST", "/graphql/", "ZM_AUTH_TOKEN=fake-token-2", bodyPayload);
 
     // When
-    HttpResponse httpResponse =
-        TestUtils.sendRequest(httpRequest, simulator.getNettyChannel());
+    HttpResponse httpResponse = app.send(httpRequest);
 
     // Then
     Assertions.assertThat(httpResponse.getStatus()).isEqualTo(200);
@@ -206,10 +182,6 @@ class AddedAndRemovedNodeNotificationMoveApiIT {
     Assertions.assertThat(notifications).hasSize(0);
 
     // reset
-    ((MockFilesConfig)
-        simulator
-            .getInjector()
-            .getInstance(FilesConfig.class))
-        .setAreNotificationsEnabled(true);
+    app.mocks().setNotificationsEnabled(true);
   }
 }
