+++ core/src/test/java/com/zextras/carbonio/files/acceptance/AddedNodeNotificationCopyApiIT.java
@@ -2,21 +2,14 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package com.zextras.carbonio.files.api.notifications;
+package com.zextras.carbonio.files.acceptance;
 
-import com.google.inject.Injector;
-import com.zextras.carbonio.files.Simulator;
-import com.zextras.carbonio.files.Simulator.SimulatorBuilder;
 import com.zextras.carbonio.files.TestUtils;
-import com.zextras.carbonio.files.api.utilities.DatabasePopulator;
+import com.zextras.carbonio.files.acceptance.seam.FilesTestApp;
+import com.zextras.carbonio.files.acceptance.seam.impl.GuiceNettyFilesTestAppBuilder;
 import com.zextras.carbonio.files.api.utilities.GraphqlCommandBuilder;
 import com.zextras.carbonio.files.api.utilities.entities.SimplePopulatorFolder;
-import com.zextras.carbonio.files.config.FilesConfig;
 import com.zextras.carbonio.files.dal.dao.ebean.ACL;
-import com.zextras.carbonio.files.dal.repositories.interfaces.FileVersionRepository;
-import com.zextras.carbonio.files.dal.repositories.interfaces.LinkRepository;
-import com.zextras.carbonio.files.dal.repositories.interfaces.NodeRepository;
-import com.zextras.carbonio.files.utilities.MockFilesConfig;
 import com.zextras.carbonio.files.utilities.http.HttpRequest;
 import com.zextras.carbonio.files.utilities.http.HttpResponse;
 import org.assertj.core.api.Assertions;
@@ -28,19 +21,14 @@ import org.junit.jupiter.api.Test;
 import java.util.List;
 import java.util.Map;
 
-class AddedNodeNotificationMoveApiIT {
+class AddedNodeNotificationCopyApiIT {
 
-  static Simulator simulator;
-  static NodeRepository nodeRepository;
-  static FileVersionRepository fileVersionRepository;
-  static LinkRepository linkRepository;
-  static MockFilesConfig mockConfig;
+  static FilesTestApp app;
 
   @BeforeAll
   static void init() {
-    simulator =
-        SimulatorBuilder.aSimulator()
-            .init()
+    app =
+        GuiceNettyFilesTestAppBuilder.aFilesTestApp()
             .withDatabase()
             .withServiceDiscover()
             .withUserManagement(
@@ -49,31 +37,25 @@ class AddedNodeNotificationMoveApiIT {
                     "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa",
                     "fake-token-2",
                     "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaab"))
-            .build()
-            .start();
-
-    final Injector injector = simulator.getInjector();
-    nodeRepository = injector.getInstance(NodeRepository.class);
-    fileVersionRepository = injector.getInstance(FileVersionRepository.class);
-    linkRepository = injector.getInstance(LinkRepository.class);
-    mockConfig = (MockFilesConfig) injector.getInstance(FilesConfig.class);
+            .build();
   }
 
   @AfterEach
   void cleanUp() {
-    simulator.resetDatabase();
-    simulator.reinitializeMocks();
-    mockConfig.setAreNotificationsEnabled(true);
+    app.backdoor().resetDatabase();
+    app.mocks().reset();
+    app.mocks().setNotificationsEnabled(true);
   }
 
   @AfterAll
   static void cleanUpAll() {
-    simulator.stopAll();
+    app.close();
   }
 
   private void createBaseScenario() {
     // Create base folder
-    DatabasePopulator.aNodePopulator(simulator.getInjector())
+    app.backdoor()
+        .populator()
         .addNode(
             new SimplePopulatorFolder(
                 "00000000-0000-0000-0000-000000000000", "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa", "folder"));
@@ -90,7 +72,7 @@ class AddedNodeNotificationMoveApiIT {
     HttpRequest httpRequest =
         HttpRequest.of("POST", "/graphql/", "ZM_AUTH_TOKEN=fake-token", bodyPayload);
 
-    TestUtils.sendRequest(httpRequest, simulator.getNettyChannel());
+    app.send(httpRequest);
 
     // Create a node on local root
     bodyPayload =
@@ -103,16 +85,16 @@ class AddedNodeNotificationMoveApiIT {
     httpRequest =
         HttpRequest.of("POST", "/graphql/", "ZM_AUTH_TOKEN=fake-token", bodyPayload);
 
-    HttpResponse httpResponse = TestUtils.sendRequest(httpRequest, simulator.getNettyChannel());
+    HttpResponse httpResponse = app.send(httpRequest);
 
     Map<String, Object> page =
         TestUtils.jsonResponseToMap(httpResponse.getBodyPayload(), "createFolder");
 
     String nodeId = (String) page.get("id");
 
-    // Move node inside of shared folder
+    // Copy node inside of shared folder
     bodyPayload =
-        GraphqlCommandBuilder.aMutationBuilder("moveNodes")
+        GraphqlCommandBuilder.aMutationBuilder("copyNodes")
             .withListOfStrings("node_ids", List.of(nodeId).toArray(new String[0]))
             .withString("destination_id", "00000000-0000-0000-0000-000000000000")
             .withWantedResultFormat("{ id }")
@@ -121,11 +103,11 @@ class AddedNodeNotificationMoveApiIT {
     httpRequest =
         HttpRequest.of("POST", "/graphql/", "ZM_AUTH_TOKEN=fake-token", bodyPayload);
 
-    TestUtils.sendRequest(httpRequest, simulator.getNettyChannel());
+    app.send(httpRequest);
   }
 
   @Test
-  void givenANodeMoveOnASharedDirectoryItShouldCreateANotificationForTheUsersItHasBeenSharedWith() {
+  void givenANodeCopyOnASharedDirectoryItShouldCreateANotificationForTheUsersItHasBeenSharedWith() {
     // Given
     createBaseScenario();
 
@@ -139,8 +121,7 @@ class AddedNodeNotificationMoveApiIT {
         HttpRequest.of("POST", "/graphql/", "ZM_AUTH_TOKEN=fake-token-2", bodyPayload);
 
     // When
-    HttpResponse httpResponse =
-        TestUtils.sendRequest(httpRequest, simulator.getNettyChannel());
+    HttpResponse httpResponse = app.send(httpRequest);
 
     // Then
     Assertions.assertThat(httpResponse.getStatus()).isEqualTo(200);
@@ -154,13 +135,9 @@ class AddedNodeNotificationMoveApiIT {
   }
 
   @Test
-  void givenANodeMoveOnASharedDirectoryAndDisabledNotificationsNoNotificationShouldBeSavedOrReturned() {
+  void givenANodeCopyOnASharedDirectoryIAndDisabledNotificationsNoNotificationShouldBeSavedOrReturned() {
     // Given
-    ((MockFilesConfig)
-        simulator
-            .getInjector()
-            .getInstance(FilesConfig.class))
-        .setAreNotificationsEnabled(false);
+    app.mocks().setNotificationsEnabled(false);
     createBaseScenario();
 
     String bodyPayload =
@@ -173,8 +150,7 @@ class AddedNodeNotificationMoveApiIT {
         HttpRequest.of("POST", "/graphql/", "ZM_AUTH_TOKEN=fake-token-2", bodyPayload);
 
     // When
-    HttpResponse httpResponse =
-        TestUtils.sendRequest(httpRequest, simulator.getNettyChannel());
+    HttpResponse httpResponse = app.send(httpRequest);
 
     // Then
     Assertions.assertThat(httpResponse.getStatus()).isEqualTo(200);
@@ -184,13 +160,9 @@ class AddedNodeNotificationMoveApiIT {
 
     final List<Map<String, Object>> notifications = (List<Map<String, Object>>) page.get("notifications");
 
-    Assertions.assertThat(notifications).hasSize(0);
+    Assertions.assertThat(notifications).hasSize(0); // One newShare and one addedNode
 
     //reset
-    ((MockFilesConfig)
-        simulator
-            .getInjector()
-            .getInstance(FilesConfig.class))
-        .setAreNotificationsEnabled(true);
+    app.mocks().setNotificationsEnabled(true);
   }
 }
