+++ core/src/test/java/com/zextras/carbonio/files/acceptance/RemovedNodeNotificationTrashApiIT.java
@@ -2,21 +2,14 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package com.zextras.carbonio.files.api.notifications;
+package com.zextras.carbonio.files.acceptance;
 
-import com.google.inject.Injector;
-import com.zextras.carbonio.files.Simulator;
-import com.zextras.carbonio.files.Simulator.SimulatorBuilder;
 import com.zextras.carbonio.files.TestUtils;
-import com.zextras.carbonio.files.api.utilities.DatabasePopulator;
+import com.zextras.carbonio.files.acceptance.seam.FilesTestApp;
+import com.zextras.carbonio.files.acceptance.seam.impl.GuiceNettyFilesTestAppBuilder;
 import com.zextras.carbonio.files.api.utilities.GraphqlCommandBuilder;
 import com.zextras.carbonio.files.api.utilities.entities.SimplePopulatorFolder;
-import com.zextras.carbonio.files.config.FilesConfig;
 import com.zextras.carbonio.files.dal.dao.ebean.ACL;
-import com.zextras.carbonio.files.dal.repositories.interfaces.FileVersionRepository;
-import com.zextras.carbonio.files.dal.repositories.interfaces.LinkRepository;
-import com.zextras.carbonio.files.dal.repositories.interfaces.NodeRepository;
-import com.zextras.carbonio.files.utilities.MockFilesConfig;
 import com.zextras.carbonio.files.utilities.http.HttpRequest;
 import com.zextras.carbonio.files.utilities.http.HttpResponse;
 import org.assertj.core.api.Assertions;
@@ -30,17 +23,12 @@ import java.util.Map;
 
 class RemovedNodeNotificationTrashApiIT {
 
-  static Simulator simulator;
-  static NodeRepository nodeRepository;
-  static FileVersionRepository fileVersionRepository;
-  static LinkRepository linkRepository;
-  static MockFilesConfig mockConfig;
+  static FilesTestApp app;
 
   @BeforeAll
   static void init() {
-    simulator =
-        SimulatorBuilder.aSimulator()
-            .init()
+    app =
+        GuiceNettyFilesTestAppBuilder.aFilesTestApp()
             .withDatabase()
             .withServiceDiscover()
             .withUserManagement(
@@ -49,31 +37,25 @@ class RemovedNodeNotificationTrashApiIT {
                     "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa",
                     "fake-token-2",
                     "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaab"))
-            .build()
-            .start();
-
-    final Injector injector = simulator.getInjector();
-    nodeRepository = injector.getInstance(NodeRepository.class);
-    fileVersionRepository = injector.getInstance(FileVersionRepository.class);
-    linkRepository = injector.getInstance(LinkRepository.class);
-    mockConfig = (MockFilesConfig) injector.getInstance(FilesConfig.class);
+            .build();
   }
 
   @AfterEach
   void cleanUp() {
-    simulator.resetDatabase();
-    simulator.reinitializeMocks();
-    mockConfig.setAreNotificationsEnabled(true);
+    app.backdoor().resetDatabase();
+    app.mocks().reset();
+    app.mocks().setNotificationsEnabled(true);
   }
 
   @AfterAll
   static void cleanUpAll() {
-    simulator.stopAll();
+    app.close();
   }
 
   private void createBaseScenario() {
     // Create the folder to be shared
-    DatabasePopulator.aNodePopulator(simulator.getInjector())
+    app.backdoor()
+        .populator()
         .addNode(
             new SimplePopulatorFolder(
                 "00000000-0000-0000-0000-000000000000", "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa", "folder"));
@@ -89,8 +71,7 @@ class RemovedNodeNotificationTrashApiIT {
     HttpRequest httpRequest =
         HttpRequest.of("POST", "/graphql/", "ZM_AUTH_TOKEN=fake-token", bodyPayload);
 
-    HttpResponse httpResponse =
-        TestUtils.sendRequest(httpRequest, simulator.getNettyChannel());
+    HttpResponse httpResponse = app.send(httpRequest);
 
     Map<String, Object> page =
         TestUtils.jsonResponseToMap(httpResponse.getBodyPayload(), "createFolder");
@@ -109,7 +90,7 @@ class RemovedNodeNotificationTrashApiIT {
     httpRequest =
         HttpRequest.of("POST", "/graphql/", "ZM_AUTH_TOKEN=fake-token", bodyPayload);
 
-    TestUtils.sendRequest(httpRequest, simulator.getNettyChannel());
+    app.send(httpRequest);
 
     // Trash node inside of base node
     bodyPayload =
@@ -121,8 +102,7 @@ class RemovedNodeNotificationTrashApiIT {
     httpRequest =
         HttpRequest.of("POST", "/graphql/", "ZM_AUTH_TOKEN=fake-token", bodyPayload);
 
-    httpResponse =
-        TestUtils.sendRequest(httpRequest, simulator.getNettyChannel());
+    httpResponse = app.send(httpRequest);
 
     Assertions.assertThat(httpResponse.getStatus()).isEqualTo(200);
   }
@@ -142,8 +122,7 @@ class RemovedNodeNotificationTrashApiIT {
         HttpRequest.of("POST", "/graphql/", "ZM_AUTH_TOKEN=fake-token-2", bodyPayload);
 
     // When
-    HttpResponse httpResponse =
-        TestUtils.sendRequest(httpRequest, simulator.getNettyChannel());
+    HttpResponse httpResponse = app.send(httpRequest);
 
     // Then
     Assertions.assertThat(httpResponse.getStatus()).isEqualTo(200);
@@ -159,11 +138,7 @@ class RemovedNodeNotificationTrashApiIT {
   @Test
   void givenANodeRemovalByTrashingItOnASharedDirectoryAndDisabledNotificationsNoNotificationShouldBeSavedOrReturned() {
     // Given
-    ((MockFilesConfig)
-        simulator
-            .getInjector()
-            .getInstance(FilesConfig.class))
-        .setAreNotificationsEnabled(false);
+    app.mocks().setNotificationsEnabled(false);
     createBaseScenario();
 
     String bodyPayload =
@@ -176,8 +151,7 @@ class RemovedNodeNotificationTrashApiIT {
         HttpRequest.of("POST", "/graphql/", "ZM_AUTH_TOKEN=fake-token-2", bodyPayload);
 
     // When
-    HttpResponse httpResponse =
-        TestUtils.sendRequest(httpRequest, simulator.getNettyChannel());
+    HttpResponse httpResponse = app.send(httpRequest);
 
     // Then
     Assertions.assertThat(httpResponse.getStatus()).isEqualTo(200);
@@ -190,10 +164,6 @@ class RemovedNodeNotificationTrashApiIT {
     Assertions.assertThat(notifications).hasSize(0);
 
     //reset
-    ((MockFilesConfig)
-        simulator
-            .getInjector()
-            .getInstance(FilesConfig.class))
-        .setAreNotificationsEnabled(true);
+    app.mocks().setNotificationsEnabled(true);
   }
 }
